requirements.txt
PyYAML
cloudshell-core==2.1.153
-e git+git@github.com:QualiSystems/cloudshell-L1-networking-core.git@refactoring#egg=cloudshell_l1_networking_core
